Updated Graphs - Fixed Scaling Issue with Negative Means
Changed the coord_cartesian() function as the original settings made the min value 0.7 * the lowest mean. This means that for all positive means the 0 point is cutoff (see default), and adding a large negative value leads to that bar not fully shown within the graph (e.g., try: 1.03, -1.21, 0.98, 1.01; the -1.21 will be cutoff).

Originally I just removed the function so it auto scales. However, I figured you both wanted the extra padding given by the 1.2 *  largest mean value. I implemented this without any cut-offs of bars. The if statements changes the padding depending if all th means are positive, negative, or a mix.